Re-organize the shared dto lib
