+++ platform/alarms/rest/src/main/java/org/opennms/netmgt/alarmd/rest/internal/AlarmRestServiceImpl.java
@@ -68,9 +68,9 @@ import org.opennms.horizon.db.model.OnmsAcknowledgment;
 import org.opennms.horizon.db.model.OnmsAlarm;
 import org.opennms.horizon.db.model.TroubleTicketState;
 import org.opennms.horizon.db.model.mapper.AlarmMapper;
-import org.opennms.horizon.shared.dto.AlarmAckDTO;
-import org.opennms.horizon.shared.dto.AlarmCollectionDTO;
-import org.opennms.horizon.shared.dto.AlarmDTO;
+import org.opennms.horizon.shared.dto.event.AlarmAckDTO;
+import org.opennms.horizon.shared.dto.event.AlarmCollectionDTO;
+import org.opennms.horizon.shared.dto.event.AlarmDTO;
 import org.opennms.netmgt.alarmd.rest.AlarmRestService;
 import org.opennms.web.rest.support.MultivaluedMapImpl;
 import org.opennms.web.rest.support.SecurityHelper;
