+++ platform/events/rest/src/main/java/org/opennms/netmgt/events/rest/internal/EventRestServiceImpl.java
@@ -68,8 +68,8 @@ import org.opennms.horizon.db.model.mapper.EventMapper;
 import org.opennms.horizon.events.api.EventForwarder;
 import org.opennms.horizon.events.xml.Event;
 import org.opennms.horizon.events.xml.Parm;
-import org.opennms.horizon.shared.dto.EventCollectionDTO;
-import org.opennms.horizon.shared.dto.EventDTO;
+import org.opennms.horizon.shared.dto.event.EventCollectionDTO;
+import org.opennms.horizon.shared.dto.event.EventDTO;
 import org.opennms.netmgt.events.rest.EventRestService;
 import org.opennms.netmgt.events.rest.ISO8601DateEditor;
 import org.opennms.netmgt.events.rest.StringXmlCalendarPropertyEditor;
