+++ platform/inventory/device/src/main/java/org/opennms/horizon/inventory/device/rest/api/DeviceRestAPI.java
@@ -37,8 +37,8 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
-import org.opennms.horizon.shared.dto.DeviceCollectionDTO;
-import org.opennms.horizon.shared.dto.DeviceDTO;
+import org.opennms.horizon.shared.dto.device.DeviceCollectionDTO;
+import org.opennms.horizon.shared.dto.device.DeviceDTO;
 
 @Path("/devices")
 public interface DeviceRestAPI {
