+++ rest-server/src/main/java/org/opennms/horizon/server/service/EventService.java
@@ -28,8 +28,8 @@
 
 package org.opennms.horizon.server.service;
 
-import org.opennms.horizon.shared.dto.EventCollectionDTO;
-import org.opennms.horizon.shared.dto.EventDTO;
+import org.opennms.horizon.shared.dto.event.EventCollectionDTO;
+import org.opennms.horizon.shared.dto.event.EventDTO;
 import org.springframework.stereotype.Service;
 
 import io.leangen.graphql.annotations.GraphQLEnvironment;
