+++ shared-lib/dto-model/src/main/java/org/opennms/horizon/shared/dto/event/AlarmSummaryDTO.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.horizon.shared.dto;
+package org.opennms.horizon.shared.dto.event;
 
 import java.beans.Transient;
 import java.util.Objects;
