+++ shared-lib/dto-model/src/main/java/org/opennms/horizon/shared/dto/event/ReductionKeyMemoDTO.java
@@ -26,23 +26,32 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.horizon.shared.dto;
+package org.opennms.horizon.shared.dto.event;
 
-import java.util.Collection;
-import java.util.List;
+import java.util.Objects;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
+public class ReductionKeyMemoDTO extends MemoDTO {
+    private String reductionKey;
 
-public class AlarmCollectionDTO extends AbstractCollectionDTO<AlarmDTO> {
-    public AlarmCollectionDTO() {
-    }
-    public AlarmCollectionDTO(final Collection<? extends AlarmDTO> alarms) {
-        objects.addAll(alarms);
-    }
-    @JsonProperty("alarms")
-    public List<AlarmDTO> getAlarms() {
-        return objects;
+    public String getReductionKey() {
+        return reductionKey;
     }
 
+    public void setReductionKey(String reductionKey) {
+        this.reductionKey = reductionKey;
+    }
 
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        if (!super.equals(o)) return false;
+        ReductionKeyMemoDTO that = (ReductionKeyMemoDTO) o;
+        return Objects.equals(reductionKey, that.reductionKey);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(super.hashCode(), reductionKey);
+    }
 }
